% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap function Modification}
\usage{
bootstrap(x, B = 1000, statistic, sample_size = length(x))
}
\description{
This bootstrap function will provide you always an object of class 'bootstrap'
}
\examples{
set.seed(566)
class(Bootstrap(rnorm(100), statistic = mean))
}
