% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{Plot Bootstrap}
\usage{
\method{plot}{bootstrap}(x, main = "Bootstrap estimates", pch = 19, cex = 1, horizontal = F)
}
\arguments{
\item{x}{is the output-list returned from the Bootstrap function}

\item{main}{represent the title of the final plot. By default it will be "Bootstrap estimates"}

\item{pch}{shape of the outliers dots. By  default is 19}

\item{cex}{size of the outliers dots. By default is 1}

\item{horizontal}{is the orientation of the box-plot. By default is F}
}
\value{
A \code{boxplot} of the bootstrap estimates.
}
\description{
It is a method of the original plot function made for the bootstrap class. It will return a box-plot.
}
\examples{
boot <- Bootstrap(rnorm(100), statistic = var)
plot(boot, main = "Test", cex=2, horizontal = T)
}
\author{
Group-E
}
