% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{plot.bootstrap.v}
\alias{plot.bootstrap.v}
\title{Plot Boot V}
\usage{
\method{plot}{bootstrap.v}(x, main = "Bootstrap estimates", pch = 19, cex = 1, horizontal = F)
}
\description{
Provide an horizontal box-plot of the class bootstrap
}
\examples{
plot(boot, main = "Test", cex=2, horizontal = T)
}
