% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{print.summary.bootstrap}
\alias{print.summary.bootstrap}
\title{Print Summary Bootstrap}
\usage{
\method{print}{summary.bootstrap}(x, digits, ...)
}
\arguments{
\item{x}{is the output-list returned from the summary.bootstrap function}

\item{digits}{are the decimals points desired in the printed output. By default digits = 2
@param ... other parameters for the 'print.summary.bootstrap' function}
}
\value{
A \code{print} of the main summary information from the Bootstrap function with the modified digits arguments. This \code{print} complementary of the first print summary.bootstrap function.
}
\description{
This print.summary.bootstrap function is a method of the basic print function. It will print a summary regarding the summary.bootstrap class generated by the summary.bootstrap function.
}
\examples{
boot <- Bootstrap(rnorm(100), statistic = var)
print(summary(boot), digits=5)
}
\author{
Group-E
}
