% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{summary.bootstrap.mod}
\alias{summary.bootstrap.mod}
\title{Summary Bootstrap Mod}
\usage{
\method{summary}{bootstrap.mod}(x, level = 0.95)
}
\description{
Modification of summary.bootstrap function to obtain "summary(boot, level = 1.2)" and "summary(boot, level = .8)"
}
\examples{
summary(boot, level = 1.2)
}
