% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{summary.bootstrap}
\alias{summary.bootstrap}
\title{Summary Bootstrap}
\usage{
\method{summary}{bootstrap}(x, level = 0.95)
}
\description{
Return an object of class summary.bootstrap that will represent a summary of the bootstrap class.
}
\examples{
summary(boot)
}
