% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{print.bootstrap}
\alias{print.bootstrap}
\title{Bootstrap Print Modification}
\usage{
\method{print}{bootstrap}(x)
}
\arguments{
\item{x}{is the output-list returned from the Bootstrap function.}
}
\value{
A \code{print} of relevant information provided by the Bootstrap function.
}
\description{
This print.bootstrap function is a method of the basic print function. The function will print information regarding the bootstrap class generated by the Bootstrap function.
}
\examples{
boot2 <- Bootstrap(matrix(rnorm(200),nc=2), statistic = var)
print(boot2)
}
\author{
Group-E
}
