% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{summary.bootstrap}
\alias{summary.bootstrap}
\title{Summary Bootstrap}
\usage{
\method{summary}{bootstrap}(x, level, digits, ...)
}
\arguments{
\item{x}{is the output-list returned from the Bootstrap function}

\item{level}{represent the confidence level of 1-alpha. By default level = 0.95}

\item{digits}{are the decimals points desired in the printed output. By default digits = 2}

\item{...}{other parameters for the 'summary.bootstrap' function}
}
\value{
A \code{print} of the main summary information from the Bootstrap function. It also return a \code{list} composed as follow:
\describe{
    \item{summary}{Will return a \code{list} composed of the basic summary function output applied on the theta_star output}
    \item{CI}{It will return a \code{list} with the lower and upper limit of the confidence interval}
    \item{sto}{The theta_hat from the Bootstrap function}
    \item{coerced_to_a_vector}{Returning a True or Flase regarding the initial sample if it was coerced to a vector during the usage of Bootstrap function}
    \item{level}{Represent the confidence level of 1-alpha called as a parameter}
    \item{stat}{Return the name of the statistic used thruough the Bootstrap function}
    \item{B}{Number of bootrsapped samples thourgh the Bootstrap function}
}
}
\description{
This summary.bootstrap function is a method of the basic summary function. It will print a summary regarding the bootstrap class generated by the Bootstrap function. The returned values will be of class summary.bootstrap.
}
\examples{
boot <- Bootstrap(rnorm(100), statistic = var)
summary(boot, level = 0.90)
}
\author{
Group-E
}
